refactor : 키관리 관련 컴포넌트명 변경
